requirements.txt
google-genai>=1.0.0
langchain>=0.3.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
